Everything Completed!
Finished 100%, Looks just like the mock-ups provided by the instructor.
The last part was a hard challenge when I had to figure out how to
position the radio buttons and checkboxes, I did a lot of research and
peer-to-peer reviews and I realised that I should use generated content
to layout and style the radio/checkboxes.